backend/requirements.txt
Django==4.2.7
gunicorn==21.2.0
psycopg2-binary==2.9.1
django-cors-headers==4.3.1
djangorestframework
djangorestframework-simplejwt
asgiref
PyJWT
pytz
sqlparse
python-dotenv
drf-yasg
pytest
pytest-django==4.5.2
factory-boy==3.3.0
pytest-factoryboy==2.5.0
pytest-cov==4.0.0
Pillow==10.3.0
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
daphne==4.0.0